refactor(web): move platform detection to utils
