+++ web/app/(commonLayout)/layout.tsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import Zendesk from '@/app/components/base/zendesk'
 import MaintenanceNotice from '@/app/components/header/maintenance-notice'
 import MainNavLayout from '@/app/components/main-nav/layout'
-import { getPlatformFromUserAgent } from '@/app/components/main-nav/platform'
 import { headers } from '@/next/headers'
+import { getPlatformFromUserAgent } from '@/utils/platform'
 import { CommonLayoutGlobalMounts } from './global-mounts'
 import { ConsoleContextProviders, ConsoleRuntimeProviders } from './providers'
