+++ web/utils/platform.ts
